+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## `7.0.0-next.202108121907`
+
+- Enhancement: Flattened the default profiles structure created by the `config init` command.
+- Breaking: Split up authToken property in team config into tokenType and tokenValue properties to be consistent with Zowe v1 profiles.
+
 ## `7.0.0-next.202107131230`
 
 - Enhancement: Adds the `config auto-init` command, allowing users to automatically generate a configuration using information stored in conformant installed plugins and the API Mediation Layer.